BackEnd/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-multipart==0.0.9
minio==7.2.15
opencv-python==4.10.0.84
numpy==2.1.3
ultralytics==8.3.181
torch==2.11.0
